major-master/requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==3.2.16
django-admin-multiple-choice-list-filter==0.1.1
django-appconf==1.0.5
django-cors-headers==3.13.0
django-imagekit==4.1.0
django-import-export==3.0.2
django-js-asset==2.0.0
django-more-admin-filters==1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
gunicorn==20.1.0
install==1.3.5
MarkupPy==1.14
numpy==1.25.1
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.3.5
pilkit==2.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pyExcelerator==0.6.4.1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
tzdata==2023.3
whitenoise==6.3.0
xlrd==2.0.1
XlsxWriter==3.1.0
xlwt==1.3.0
